logger threading fix
Apparently pthread_cond_signal (and possibly pthread_cond_broadcast) has to be
send before pthread_mutex_unlock. It makes some sense as it prevents the
waiting thread from kicking in the very moment the signal is send and let's the
function finish execution, but I'm not sure of the ordering significance in
this scenario. Anyway, it appears to be working.